Refactored acf section code
